% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namesexdata.R
\docType{data}
\name{brazilian_pols}
\alias{brazilian_pols}
\title{Brazilian Politician Dataset}
\format{A data frame with 76542 observations and 8 variables}
\source{
Brazilian Election Data,
  \url{http://www.tse.jus.br/hotSites/pesquisas-eleitorais/candidatos.html}
}
\usage{
brazilian_pols
}
\description{
A dataset containing 76,542 Brazilian first names and how often each name
is used for male or female candidates in Brazil's 1998-2014 general and
municipal elections.
}
\author{
Luke Sonnet
}
\keyword{datasets}

