% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namesexdata.R
\docType{data}
\name{indian_pols}
\alias{indian_pols}
\title{Indian Politician Dataset}
\format{A data frame with observations and 4 variables}
\source{
MyNeta,
  \url{http://www.myneta.info/}
}
\usage{
indian_pols
}
\description{
A data set containing the names of Indian politicians.
}
\author{
Gaurav Sood
}
\keyword{datasets}

