% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namesexdata.R
\docType{data}
\name{brazilian_pols}
\alias{brazilian_pols}
\title{Brazilian Politician Dataset}
\format{A data frame with 74650 observations and 6 variables}
\source{
Brazilian Muncipal Election Data,
  \url{http://www.tse.jus.br/hotSites/pesquisas-eleitorais/candidatos.html}
}
\usage{
brazilian_pols
}
\description{
A dataset containing 74,000 Brazilian first names and how often each name 
is used for a male or female candidates in the 2000, 2004, 2008, and 2012 
municipal elections in Brazil.
}
\author{
Luke Sonnet
}
\keyword{datasets}

